adding comments
goals:
add points
clean up title screen
add hard, normal, easy modes
split code into different files
make comments more professional
if I can shift the fruit to be in positions by 20 the eat detection should be changed to hit detection


(from previous)
When the body grows it is set to grow as if the snake was going down. I have set bunch of if statements which should grow the body depending the end rectangle direction, but for some reason it does not work. This may have worked in the past but not sure, The current fix is to create box outside the range of the game which creates a delayed effect on the growth.